changed web3 library to use ether-rs
using ethers-rs to make contract call and query since web3 wasn't sufficient
had to refactor its functions and types
added xauto implementation
fixed issue with getting layer2 assets

+++ src/utilities/layer2_assets.rs
@@ -12,7 +12,7 @@ pub struct layer2_asset {
     pub protocol_name: String,
     pub protocol_address: String,
     pub protocol_abi: String,
-    pub network: i32,
+    pub network: u64,
     pub decimals: i32,
     pub widthdraw_decimals: i32,
     pub ppfs_method: String,
@@ -196,7 +196,7 @@ pub fn get_layer2_assets() -> [layer2_asset; 11] {
 }
 
 
-pub async fn filter_token(token_name: String, network: i32, protocol: String) -> layer2_asset {
+pub async fn filter_token(token_name: String, network: u64, protocol: String) -> layer2_asset {
     let result: [layer2_asset; 11] = get_layer2_assets();
     let asset_result = result
         .into_iter()
